back/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
Pillow==11.3.0
gunicorn==20.1.0
psycopg2-binary==2.9.10
djoser==2.3.3
djangorestframework-simplejwt==5.5.1